pyproject.toml
[tool.poetry]
name = "langgraph-demo"
version = "0.1.0"
description = ""
authors = ["neelcohere <neel@cohere.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-cohere = "^0.1.9"
langgraph = "^0.1.8"
langsmith = "^0.1.85"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
